\name{cor.confidence.interval}
\alias{cor.confidence.interval}
\title{This calcultets the confidence interval}
\usage{
cor.confidence.interval(r1, r2 = NA, n, signif = 0.05)
}
\arguments{
  \item{r1,r2}{correltion coefficients of the two paired
  samples}

  \item{n}{length (size) of the two paired samples}

  \item{signif}{significance}
}
\description{
This calcultets the confidence interval
}
\seealso{
\url{http://onlinestatbook.com/lms/estimation/correlation_ci.html}
or \url{http://en.wikipedia.org/wiki/Fisher_transformation}
}

