\name{cor.diff.test}
\alias{cor.diff.test}
\title{cor.diff.test}
\usage{
cor.diff.test(r1, r2, n1, n2, alternative = c("two.sided", "less", "greater"))
}
\description{
cor.diff.test description
}
\references{
\url{http://stackoverflow.com/questions/14519006/significance-test-on-the-difference-of-two-correlation-coefficient}
}

