\name{corplotlags}
\alias{corplotlags}
\title{Function which plots the correlation among observed and generated variables for more lags}
\usage{
corplotlags(x = Tn_mes, y = Tn_gen, lag = c(0, 1, 4),
  return.just.data.frame = FALSE, xlab = "observed", ylab = "generated",
  title = "Spatial Correlation", ...)
}
\description{
Function which plots the correlation among observed and
generated variables for more lags
}
\seealso{
\code{\link{corplotlag}}
}

