\name{fun_factor}
\alias{fun_factor}
\title{Applies a function to a data frame according to a FACTOR field.}
\usage{
fun_factor(data, factor = NA, fun, returns.data.frame = TRUE, ...)
}
\arguments{
  \item{dataset}{}

  \item{factor}{field of \code{data} which is a factor.}
}
\description{
Applies a function to a data frame according to a FACTOR
field.
}

