\name{logcor.plot}
\alias{logcor.plot}
\title{logcor.plot}
\usage{
logcor.plot(prec = prec_gen, Tx = Tx_gen, Tn = Tn_gen,
  station = station, origin = "1961-01-01", valmin = 0.5, xlab = "xlab",
  ylab = "ylab", title = "tilte", ...)
}
\arguments{
  \item{prec_gen}{list of precipitation data frames}

  \item{Tx_gen,Tn_gen}{list of daily maximim and minimum
  temperature data frames}
}
\description{
Correlation between log-precipitation and daily temperature
}

