\name{fun_factor}
\alias{fun_factor}
\title{Applies a function to a data frame according to a FACTOR field.}
\usage{
fun_factor(data, factor = NA, fun, returns.data.frame = TRUE, ...)
}
\arguments{
  \item{data}{input dataset as a data.frame class}

  \item{factor}{field of \code{data} which is a factor.}

  \item{fun}{function}

  \item{returns.data.frame}{logical value. Default is
  \code{TRUE}.}

  \item{...}{further arguments or \code{fun}}
}
\description{
Applies a function to a data frame according to a FACTOR
field.
}

