\name{temperature.wetdry.barplot}
\alias{temperature.wetdry.barplot}
\title{temperature.wetdry.barplot}
\usage{
temperature.wetdry.barplot(prec = prec_gen, Tx = Tx_gen, Tn = Tn_gen,
  station, origin = "1961-01-01", valmin = 0.5, xlab = "xlab",
  ylab = "ylab", title = "title", variable, ...)
}
\description{
Bar plot for temperature differantiating "dry" or "wet"
days.
}

